organize methods better
