+++ cg_deployment/models.py
@@ -1,5 +1,5 @@
 import datetime
-from typing import Optional, List, Any
+from typing import Any, List, Optional
 
 from pydantic import BaseModel
